fixed task 0

+++ search_algorithms/0-linear.c
@@ -20,10 +20,10 @@ int linear_search(int *array, size_t size, int value)
 
 	while (i < size)
 	{
-		printf("Value checked array[%d] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
 
 		if (array[i] == value)
-			return (i); /* return index of value found */
+			return ((int) i); /* return index of value found */
 
 		i++;
 	}
